+++ src/lib/process-item-enum.ts
@@ -1,0 +1,15 @@
+/**
+ * Process step type enum
+ */
+export enum ProcessItemTypeEnum {
+  CALL_RECEIVED = 'CALL_RECEIVED',
+  IMAGES_RECEIVED = 'IMAGES_RECEIVED',
+  KVA_REQUESTED = 'KVA_REQUESTED',
+  KVA_RECEIVED = 'KVA_RECEIVED',
+  LEASE_APPROVAL = 'LEASE_APPROVAL',
+  FLEET_APPROVAL = 'FLEET_APPROVAL',
+  INSURANCE_APPROVAL = 'INSURANCE_APPROVAL',
+  WORKSHOP_APPOINTMENT = 'WORKSHOP_APPOINTMENT',
+  REPAIRED = 'REPAIRED',
+  PROCESS_COMPLETED = 'PROCESS_COMPLETED',
+}
